fix(osrs-hiscores): fix an error when trying to load osrs players activities
The issue was the Jagex removed Clues from the activities therefore there are no longer 3 different
activities to load, causing the crash. This is a quick patch, I've heard they plan to add all clue
tiers to the activities hiscores, so when that happens the config will need to be updated once
again.